Componentization Pure
